Update WGMenu to inherit from our QWidget backed menu rather than the QML menu. Automatically convert MenuItem children of WGMenu into WGActions to maintain backwards compatibility with controls already using WGMenu.qml
Update the layout manger to no longer match an action to a single menu but to every menu that is compatible.
This required the addition of the '|' character to action paths to determine a minimum path that must be matched when testing against menus.

Fixed erroneous shared action handling that was causing memory leaks.