Fix for crash when displaying methods that have no value in the UI.
Problem:
Python has objects that are both method and value so need to be displayable by the UI.
Reflection system also has objects that are only method and are not displayable by the UI.

Resolution:
Separate checks into 3 types canGet/isVaue, canSet/readonly and canInvoke/isMethod.
Added some asserts for all 3 types of check on access.
Display all items in the UI, but just put a blank for items that have no value.
NGT-1583