Fix for non-integer keys not displaying in maps.
Problem
Map keys always displayed as [0], [1], [2], even if the keys were not integers.

Solution
Set a display name separate from the property name in ReflectedPropertyItem.
E.g. property path map[0] can have a display name "[key name]", so it will display as "[key name]" instead of "[0]".
Added a test for non-integer keys to the test page.
Added an assert parent isn't null, because it was de-referencing it in the initializer list.

NGT-553, NGT-1603