Fix crash on shutdown.
Problem
It would crash on shutdown because the qml files hold references to Python objects, but Python had already been finalized.
The qml files create QtScriptObjects, which are owned globally by the scripting engine and only get deleted on shutdown.

Solution
Move the Python plugin so that it shuts down after the scripting engine shuts down. So that the scripting engine can delete Python objects before Python is finalized.

NGT-553