Fix another crash on exit in ngt-bw. Or after changing properties in the Python test panel.
Problem
The return value of SetReflectedPropertyCommand stores a strong references to objects in the CommandManager history. This causes objects to be destroyed at the wrong time during shutdown.

Solution
Do not return the object in SetReflectedPropertyCommand. Not sure how this affects the rest of 5437de4010fde0d25cdbc9c24d1a2700f1648716 though.

NGT-1547