Fix crash on exit after browsing certain objects.
Problem
After expanding newStyleObject and oldStyleObject it would assert on exit that the definition for "ConstructorTest1" had already been removed.
This is because there are two different classes called "ConstructorTest1" in the test script, but DefinitionManager cannot distinguish them because of the same type name.
E.g.
# module "Test"
class A( object ):
    class InnerClass( object ):
        pass
    pass
class B( object ):
    class InnerClass( object ): # different type with the name "Test.InnerClass"
        pass
    pass

Solution
Add an id after the type name to distinguish different types which have the same name.
Using the same method as Python's built-in id() function.
Added an assert to DefinitionManager that it is not adding 2 different definitions with the same name.