Fix for actions.xml not loading.
Problem
Copy/paste actions were not being added to the menu correctly.
But no error was logged.

Solution
Added an error logger to qt_action_manager.cpp.
The error was that the actions.xml was not loaded (empty file), adding ":/" to the front of paths seems to fix this.
Fixed similar errors in other plugins that the error logger detected.

NGT-2108