Fix for unit tests that run as a plugin.
Problem
Unit tests that run with a plugins.txt started crashing during LoadLibraryW().
Because CppUnitLite2 will initialize static variables that call "new". And then "new" accesses the global plugin context to find the memory allocator. But the global plugin context is initialized by another static variable. Depending on which static is initialized first, it may work or may not work.

Solution
Initialize the global plugin context on first access if it was not already initialized.

NGT-1949