Core NGT changes from WGS up to 05/24/2016
Added WGTableView.qml to WGControls (currently the column data model is specified in wgs_multi_edit) ideally we could interchangeably use a javascript model or a C++ model.
Modified ClassDefinition to take an unique_ptr<IClassDefinitionDetails> to indicate that it takes ownership of the value.
Added link dependency in core_qt_common to core_qt_script.